requirements.txt
pyyaml
tqdm
opencv-python
imgaug
lpips
plyfile
tensorboard
sympy
numpy
pillow
matplotlib
bidict
ipdb
imageio
imageio-ffmpeg
open3d
timm
torchmetrics
kornia
roma
imgviz
protobuf==3.19
termcolor
jupyter
segment_anything
git+https://github.com/NVlabs/nvdiffrast.git
